fix(src): Sistemati Metodi Similarity
Modificati metodi Similarity per essere adattati alla nuova
struttuara della libreria.

Aggiornata documentazione
