+++ docs/similarity_methods/index.html
@@ -52,16 +52,13 @@ The similarity measures between every pair can be _calculated using several
 properties of the network_, one of which is structural property.
 Scores based on this property can be grouped in several categories
 like **local** and **global**, and so on.
-&#34;&#34;&#34;
-from . import local_similarity
-from . import quasi_local_similarity
-from . import global_similarity</code></pre>
+&#34;&#34;&#34;</code></pre>
 </details>
 </section>
 <section>
 <h2 class="section-title" id="header-submodules">Sub-modules</h2>
 <dl>
-<dt><code class="name"><a title="cnlp.similarity_methods.global_similarity" href="global_similarity/index.html">cnlp.similarity_methods.global_similarity</a></code></dt>
+<dt><code class="name"><a title="cnlp.similarity_methods.global_similarity" href="global_similarity.html">cnlp.similarity_methods.global_similarity</a></code></dt>
 <dd>
 <div class="desc"><p>Collection of Global Similarity Methods for Link Prediction …</p></div>
 </dd>
@@ -69,7 +66,7 @@ from . import global_similarity</code></pre>
 <dd>
 <div class="desc"><p>Collection of Local Similarity Methods for Link Prediction …</p></div>
 </dd>
-<dt><code class="name"><a title="cnlp.similarity_methods.quasi_local_similarity" href="quasi_local_similarity/index.html">cnlp.similarity_methods.quasi_local_similarity</a></code></dt>
+<dt><code class="name"><a title="cnlp.similarity_methods.quasi_local_similarity" href="quasi_local_similarity.html">cnlp.similarity_methods.quasi_local_similarity</a></code></dt>
 <dd>
 <div class="desc"><p>Collection of Quasi-local Similarity Methods for Link Prediction …</p></div>
 </dd>
@@ -95,9 +92,9 @@ from . import global_similarity</code></pre>
 </li>
 <li><h3><a href="#header-submodules">Sub-modules</a></h3>
 <ul>
-<li><code><a title="cnlp.similarity_methods.global_similarity" href="global_similarity/index.html">cnlp.similarity_methods.global_similarity</a></code></li>
+<li><code><a title="cnlp.similarity_methods.global_similarity" href="global_similarity.html">cnlp.similarity_methods.global_similarity</a></code></li>
 <li><code><a title="cnlp.similarity_methods.local_similarity" href="local_similarity.html">cnlp.similarity_methods.local_similarity</a></code></li>
-<li><code><a title="cnlp.similarity_methods.quasi_local_similarity" href="quasi_local_similarity/index.html">cnlp.similarity_methods.quasi_local_similarity</a></code></li>
+<li><code><a title="cnlp.similarity_methods.quasi_local_similarity" href="quasi_local_similarity.html">cnlp.similarity_methods.quasi_local_similarity</a></code></li>
 </ul>
 </li>
 </ul>
